refactor: generic collision handling (#59)
Related to:
- #33

I have reworked our collision system to make it as generic as possible.
This implies multiple changes in the way we handle collisions:

---
No more collision storage. We store collisions as entities and assign
them a component. This allows to store any information needed for
different types of collisions.

---
Naming convention: names were all over the place and it was hard to
differenciate box collisions and soft body collisions. I have used the
following convention: `<verb><collisionType>Collisions`
For example, `DetectABABCollisions` or `ApplySoftBodyCollisions`

Collisions are split into three steps:
- Detection
- Application (if applicable, no pun intended)
- Deletion

Detection only performs the algebra to detect if two objects actually
collides and creates a new collision entity according to what happened.
Application updates the colliding components and only happens if it is
necessary. For example, with box collisions, it is up to the developer
to decide what they want to do with the collision, but with soft body
simulations we need to provide a way to update the particle positions.
Deletion simply deletes the collision entities.

Systems should be added in this order as well

---
I had to merge physics and collisions because of circular patterns
happening with the soft body collisions.
Right now there are still two namespaces Physics and Collisions

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>